[11.x] Test Improvements
- Split redundant code
